Introduced ALmixer_SuspendPlayingState(), ALmixer_ResumePlayingState, ALmixer_IsPlayingStateSuspended to make it easier to properly handle ALmixer_BeginInterruption/ALmixer_EndInterruption. These will check the currently playing status of a channel, and pause the channel when calling SuspendPlayingState. ResumePlayingState will resume all the channels that were playing when SuspendPlayingState was called. This avoids the pitfall of resuming everything where some channels may have intentionally been paused during an interruption event and you don't want to automatically resume those.
Also realized that some of the queury APIs like getting the number of channels don't need to fail when the system is in an interruption state. I originally tried to implement the above functions at the user level, but got blocked because those functions wouldn't work inside an interruption. Pause/Play will still not work in an interruption which was a consideration for adding the functions above. But these functions for the most part seem reasonable to work in interruptions. (AllocateChannels is a little weird since it does either allocation or returns the number. I'm disliking the overloaded nature of this function.)