SWIG Changes to alGetSource?f?.
Realized that I can't dynamically change the type in JavaScript as a parameter (e.g, what if they passed a number constant). This means I always have to modify a passed in array and can't make it a flat value in the alGetSourcef case.

Also, I can't figure out if/how to return a value for alGetSourcef because I don't know how to connect argout and out, nor do I understand how to say for just this void return value, allow this to return a value (but not for the other functions).

alGetSource3f() doesn't really work.

	ALmixer.alGetSource3f(alsource, ALmixer.AL_POSITION, ret_val1, ret_val2, ret_val3);
	  //Ti.API.info("position  "+ ret_array[0] + ", " +  ret_array[1] + ", " +  ret_array[2]);
	  Ti.API.info("alGetSource3f position  "+ ret_val1[0] + ", " +  ret_val2[0] + ", " +  ret_val3[0]);

What happens is that all the values are getting put in the first array (size 3), instead of one element in each array.