Javascript bootstrap: Fixed a race condition bug caused by the fact that the JS callbacks may be deferred (due to redirecting to other threads).
The problem is that to deal with memory management,
I'm keeping my own shadow table that keeps ALmixer_Data rooted and keeps mappings
to scripting language defined callback functions.
If Play is immediately called after Halt(), but the callback hasn't fired yet,
I will accidentally overwrite the callback data that is supposed to fire with
the next Play's callback data.
To avoid, I need to keep a queue of callback data which PushCallbackTable and ShiftCallbackTable provide.