Change to not call rewind in LoadStream because OpenSL ES/Android seems to have bugs doing this.
The changes are somewhat intrusive because they dance around previous workarounds. The first seems to be an Ogg Vorbis workaround for very small files that may fit in the pre decoded optimization because EOF is not immediately clear without a second read.
The second deals with a rework to not queue until play to deal with OpenAL queuing bugs I don't fully remember. But I believe all the queuing problems I had have been fixed in the implementations so I'm not as worried about changing this.

Testing of this has been light, and data access has not really been tested.