This is a lame hack to get around a really complicated problem relating to: 1) Android doesn't run the reinitialization of static variables to NULL at the top of this program. 2) When dealing with a garbage collected environment, if Quit() is called before the GC environment collects, dangling pointers will be passed to FreeData() 3) We are client of somebody else's environment and can't ensure Quit() is called after the GC environment is clean. 4) Android atexit() doesn't seem to actually trigger anything.
This will clean up as much as possible.
The linked list of data will be left alive and it will dangle after the program quits. Fortunately ALmixer doesn't null check the linked list and always recreates it on Init().
Calling alBufferData after the OpenAL device is closed may cause problems depending on implementations.

Additionally ALmixer_FreeData has been changed to no longer abort if ALmixer is not initialized or the context is not active. It will try to free for better or worse. I think there is still a bug where ALmixer is re-initialized and say GC tries to free an object. The linked list is problematic, particularly if the address got reused. A recently loaded file will be unloaded causing mysterious problems. I don't see a good way around this problem though.

Also moved my test case/hacks (done on Mac) to tests to preserve.