Windows: Waypoint check in for Windows Media Foundation backend.
This implementation is incomplete and unusable, but it compiles and runs.
Windows Media Foundation currently opens a hard coded file and decodes it. The test programs playsound and playstream work with it.

Seeking/rewinding is completely untested.

The next major piece is to replace the hardcoded file with RWops, but this looks hard because the Buffer vs. URL abstraction looks a lot harder with WMF.