Android OpenSL ES is fixed again. System now utilizes SDL_Rwops without our previous hacks. (But there are new hacks.) CMake build system fixed too.
1) If not using SDL, you must call ALmixer_Android_Init(jobject activity_context) (ALmixer_PlatformExtensions.h). This is so the RWops system can access the context to reach things.
2) RWops is limited to handling real files (that can return an 'int fd')


Note that I had to modify the SDL_android.c implementation for RWops handling slightly because it assumed it was using an SDLActivity class which provided a custom method called getContext. Since ALmixer can be embedded in any situation, I can't rely on this and the code was modified to deal with the Activity instance you provide to fetch things like the asset manager.


ALmixer_OpenSLES_Android_SetAssetManager() is now removed.