Seek fix for Windows Media Foundation backend.
Calling SetCurrentPosition would trigger an error and I would be unable to seek.
(It's not the MF_E_INVALIDREQUEST error, but the else case. Also, I see a first-chance exception message in the console.)
(With or without the Flush call).
This was discovered with the LoadStream looping feature.
It seems that if I defer calling SetCurrentPosition until the Read callback, then I can seek there.
I don't really understand the reasoning for this.
The Microsoft documentation says nothing, and Google searches continue to show that pretty much nobody else in the world uses this API.
So I changed the code to defer calling seek.
The big flaw in this fix is that I can't return a legitimate return code because I don't know yet if the seek will work.

Other details:
There was already an isSeeking flag implemented.
So I needed to save the time position and and also make a modification to the read callback to call SetCurrentPosition.
Also note that the original code does not have the isSeeking check early enough so I put in an extra early block for this.
My current theories for why I have to do this:
- It could be that SetCurrentPosition can only be called from a certain thread.
- It could be I have to wait for internal buffers to clear. e.g. Maybe I call flush, and then there is an async delay to actually flush
which I have to wait for. But I don't think I enabled async flush so I think it's supposed to be synchronous. But if this is the problem,
I don't actually know when the buffers clear, so it could be dumb luck that this fix is working right now.
- Maybe SetCurrentPosition can only be called in certain callbacks? (Documentation says nothing.)
- MediaFoundation is just plain broken (and nobody uses it so nobody knows)